Add nick colourization support
This uses weechat_info_get("nick_color_name"), so the colour hashing
should be the same as in IRC, and can be turned off via the global
weechat settings in the same manner if undesired. The fallback to
irc_nick_color_name is there for older versions of weechat, but (as I
don't have an older version handy) is untested.